Rename Tab to TabLayout
Rename Tab Component to TabLayout component
Separate Route definition from component definition
Add Route definition for TabLayout into Routes.js
