+++ src/layouts/TabLayout/index.js
@@ -1,4 +1,4 @@
 // Copyright (c) Cosmo Tech.
 // Licensed under the MIT license.
 
-export { default } from './Tab';
+export { default } from './TabLayout';
