+++ src/layouts/index.js
@@ -1,4 +1,4 @@
 // Copyright (c) Cosmo Tech.
 // Licensed under the MIT license.
 
-export { default as Tab } from './Tab';
+export { default as TabLayout } from './TabLayout';
